+++ __init__.py
@@ -1,24 +1,8 @@
-import sys
 import os
-import time
-import json
-import re
-import math
-from decimal import *
-import getpass
-import random
-import subprocess
-import signal
-import traceback
-from web import sesion
-import getJson
-import subirEdificio
-import getIds
-import getStatus
-import enter
-import run
-import clear
-import update
-import fork
-import signals
-import read
+from sisop.varios import *
+
+path = os.path.abspath(__file__)
+path = os.path.dirname(path)
+os.chdir(path)
+run('touch ' + cookieFile)
+run('touch ' + telegramFile)
