--- sisop/clear_lala.py
@@ -1,7 +1,0 @@
-#! /usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-import os
-
-def clear():
-	os.system('clear')
