--- sisop/enter.py
@@ -1,7 +1,0 @@
-#! /usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-import getpass
-
-def enter():
-	getpass.getpass('\n[Enter]')
