--- sisop/fork_lala.py
@@ -1,18 +1,0 @@
-#! /usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-import os
-from sisop.run import *
-
-def forkear(s):
-	newpid = os.fork()
-	if newpid != 0:
-		# padre
-		s.login()
-		newpid = str(newpid)
-		run('kill -SIGSTOP ' + newpid)
-		run('bg ' + newpid)
-		run('disown ' + newpid)
-	else:
-		# hijo
-		s.padre = False
